Separated handler tests into distinct areas
Given the large number of tests for the handler, we have separated these into distinct sort and ordering options and placed each into separate test classes to make it easier to find and work with the current set of tests. The handler itself will be refactored after search has been added.