Changed logic for getting apprentices to make it more efficient
The logic has been changed to select the minimal amount of sql queries to run depending on what is being returned. The number of apprentices to skip for paging and the number being returned by the first sql request is taking into account to remove the need to re-query the database or to skip the first query if the skip count is great enough.