Refactored Get Apprenticeships handler to simply classes
I've broken down the handler as it was getting way to large for a class and have taken the 3 different types of search apprenticeships and placed them in their own handlers (not mediator ones). The code is not fully refactored yet but this is a good milestone to check in the latest changes. The locations of the new files is likely to change and we still have a few more areas that could be broken down and simplified.