Changed the sort logic for status to properly sort active state
The sorting for apprentices is done via the payment status but the apprenticeship status is returned and is derived from the payment status. As such, there is a problem when we derive from the active payment status as it can be derived into two different states and so we need to also sort by start date as well to correct sort the apprentices status.